new dev

+++ src/main/java/com/example/food_web/service/impl/FoodService.java
@@ -2,6 +2,7 @@ package com.example.food_web.service.impl;
 
 import com.example.food_web.model.Food;
 import com.example.food_web.repository.IFoodRepository;
+import com.example.food_web.service.IBillService;
 import com.example.food_web.service.IFoodService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -14,6 +15,9 @@ import java.util.Optional;
 public class FoodService implements IFoodService {
     @Autowired
     public IFoodRepository iFoodRepository;
+
+    @Autowired
+    public IBillService iBillService;
     @Override
     public List<Food> findAll() {
         return iFoodRepository.findAll() ;
@@ -31,11 +35,31 @@ public class FoodService implements IFoodService {
 
     @Override
     public void delete(Long aLong) {
-        iFoodRepository.deleteById(aLong);
+        Optional<Food> foodOptional = findOne(aLong);
+//        List<Bill> billListByFoodId = iBillService.findAllBillByFoodId(aLong);
+//        List<Bill> billLists = iBillService.findAll();
+        if (foodOptional.isPresent()){
+//            billLists.removeAll(billListByFoodId);
+//            iBillService.deleteBillFood(aLong);
+            iFoodRepository.deleteById(aLong);
+        }
+
     }
 
     @Override
     public Page<Food> findAllPage(Pageable pageable) {
         return iFoodRepository.findAll(pageable);
     }
+
+    @Override
+    public Page<Food> findProductByNameAbout(String name, Pageable pageable) {
+        return iFoodRepository.findProductByNameAbout("%" + name + "%", pageable);
+    }
+
+
+    @Override
+    public Page<Food> filter(Long min, Long max, String name, Pageable pageable) {
+        return iFoodRepository.filter(min, max,"%" + name + "%", pageable);
+    }
+
 }
